requirements.txt
fast-slic==0.4.0
matplotlib==3.3.4
opencv-python==4.5.1.48
pillow==7.2.0
pycocotools==2.0.2
pykeops==1.5
sacred==0.8.2
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
sklearn==0.0
tensorboard==2.4.1
tensorboardx==2.1
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
# conda install -c anaconda cudatoolkit==10.0.130
# conda install -c anaconda cudnn==7.6.5